Move CheckoutSession mapping logic to a new CheckoutSessionMappers file.
